[#220] Refactor: png -> avif 확장자 변경
