+++ pages/login/index.tsx
@@ -5,8 +5,8 @@ import LoginButton from '@/components/common/Button/LoginButton';
 import * as C from '@/constants/SignValidate';
 import useAuth, { LoginFormValue } from '@/hooks/Auth/useAuth';
 import useRedirectByLogin from '@/hooks/Auth/useRedirectByLogin';
-import mainLogoText from '@/public/assets/images/title.png';
-import mainLogo from '@/public/assets/images/transJaws.png';
+import mainLogoText from '@/public/assets/images/title.avif';
+import mainLogo from '@/public/assets/images/transJaws.avif';
 import { fontStyle } from '@/styles/fontStyle';
 import { onMobile } from '@/styles/mediaQuery';
 import { COLORS } from '@/styles/palettes';
